Refactored code to better follow clean architecture
