+++ .env
@@ -1,1 +1,1 @@
-GEMINI_API_KEY=
+GEMINI_API_KEY=AIzaSyBSCCE-TrdgcTDNEfR6KkfK0KJCSECPNns
