+++ .idea/shelf/Uncommitted_changes_before_Checkout_at_2025-12-01_4_34_p_m___Changes_1.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Checkout_at_2025-12-01_4_34_p_m__[Changes]1" date="1764624857326" recycled="true" deleted="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Checkout_at_2025-12-01_4_34_p_m__[Changes]1/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Checkout at 2025-12-01 4:34 p.m. [Changes]" />
+</changelist>
