+++ .idea/shelf/Uncommitted_changes_before_Update_at_2025-12-01_4_34_p_m___Changes_.xml
@@ -1,0 +1,4 @@
+<changelist name="Uncommitted_changes_before_Update_at_2025-12-01_4_34_p_m__[Changes]" date="1764624873156" recycled="false" toDelete="true">
+  <option name="PATH" value="$PROJECT_DIR$/.idea/shelf/Uncommitted_changes_before_Update_at_2025-12-01_4_34_p_m__[Changes]/shelved.patch" />
+  <option name="DESCRIPTION" value="Uncommitted changes before Update at 2025-12-01 4:34 p.m. [Changes]" />
+</changelist>
